Move lambert_w0 import and test to its own file
